+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- The mypy plugin being broken as of mypy 0.930.
+  (Namely by these changes <https://github.com/python/mypy/pull/11332>)
 
 ## [0.4.0] - 2021-12-17
 ### Added